+++ src/index.js
@@ -3,36 +3,14 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { createBrowserRouter, Outlet, RouterProvider } from 'react-router-dom';
-import Body from './Body';
-import AboutPage from './About';
-import ContactPage from './Contact';
-import Header from './Header';
-import Footer from './Footer';
 
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Body />
-  },
-  {
-    path: "/about",
-    element: <AboutPage />
-  },
-  {
-    path: "/contact",
-    element: <ContactPage />
-  }
-])
+
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Header />
-    <RouterProvider router={router}>
-      <App />
-    </RouterProvider>
-    <Footer />
+    <App/>
   </React.StrictMode>
 );
